fix delete

+++ src/services/profiles_services.py
@@ -147,13 +147,16 @@ def update_profile(profile_id, data):
 def delete_profile(profile_id, user_id, device_id):
     profile = Profiles.query.get(profile_id)
     if profile:
-        hma_services.delete_browser_profile(profile.hma_profile_id, user_id, device_id)
-        Events.query.filter_by(profile_id=profile_id).delete()
-        Events.query.filter_by(profile_id_interact=profile_id).delete()
-        Posts.query.filter_by(profile_id=profile_id).delete()
-        db.session.delete(profile)
-        db.session.flush()
-        return True
+        delete_status = hma_services.delete_browser_profile(
+            profile.hma_profile_id, user_id, device_id
+        )
+        if delete_status:
+            Events.query.filter_by(profile_id=profile_id).delete()
+            Events.query.filter_by(profile_id_interact=profile_id).delete()
+            Posts.query.filter_by(profile_id=profile_id).delete()
+            db.session.delete(profile)
+            db.session.flush()
+            return True
     return False
